Move websocket in a subfolder
